Fix provider type casting and nolint placement
Corrects type handling in provider clients and fixes linter directive
placement for proper bodyclose analysis.

- provider/gitlab: Remove redundant type conversion (project.ID is
  already int64, matches c.projectID type)
- provider/notion: Remove redundant Status type cast (opts.Status
  is already provider.Status, matches function parameter)
- provider/trello: Move nolint comment to same line as http.Do call
  so bodyclose linter recognizes the directive

+++ prototype/internal/provider/gitlab/client.go
@@ -85,7 +85,7 @@ func (c *Client) getProjectID(ctx context.Context) (int64, error) {
 		return 0, wrapAPIError(err)
 	}
 
-	c.projectID = int64(project.ID)
+	c.projectID = project.ID
 	return c.projectID, nil
 }
